set Disconnecting status during disconnect
Since disconnect will take over a second for a FTMS trainer, since the
disconnect procedure evolves sending a reset command and waiting for at least
a second to get a success response, the User might think the UI is frozen.
So we set status to disconnecting which will set the connection indicator
from green to yellow giving the User indication that something is happening.